Refactor import paths for createCSVfile and add createCSVfile class for CSV output generation
