+++ src/gnssmultipath/utils/__init__.py
@@ -1,4 +1,5 @@
 from .cddis_downloader import CDDISDownloader
+from .createCSVfile import createCSVfile
 from .PickleHandler import PickleHandler
 from .StatisticalAnalysis import StatisticalAnalysis
 from .writeOutputFile import writeOutputFile
